+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## unreleased
 
+- fix: Deserialize envelope with header and item #620
 - fix: Set LogLevel in startWithConfigureOptions #613
 
 ## 5.1.10-beta.0